% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/are_prodigal_options.R
\name{are_prodigal_options}
\alias{are_prodigal_options}
\title{Determine if the Prodigal options are valid}
\usage{
are_prodigal_options(prodigal_options)
}
\arguments{
\item{prodigal_options}{options to run Prodigal,
as created by \link{create_prodigal_options}}
}
\description{
Determine if the Prodigal options are valid
}
