% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_prodigal_files.R
\name{create_prodigal_files}
\alias{create_prodigal_files}
\title{Create the Prodigal output files, by running Prodigal}
\usage{
create_prodigal_files(
  fasta_filename,
  folder_name = get_default_prodigal_folder(),
  temp_folder_name = tempfile(pattern = "prodigal_")
)
}
\arguments{
\item{fasta_filename}{path to a FASTA file}

\item{folder_name}{superfolder of Prodigal.
The superfolder's name is \code{/home/[user_name]/.local/share}
by default, as can be obtained by
\link{get_default_prodigal_folder}}

\item{temp_folder_name}{path of a temporary folder.
The folder does not need to exist.
Files that are out in this folder are not automatically
deleted, which is not a problem, as the default
path given by \link{tempdir} is automatically cleaned
by the operating system}
}
\value{
full path to the files created
}
\description{
Create the Prodigal output files, by running Prodigal
}
\examples{
library(testthat)

if (1 == 2 && is_prodigal_installed()) {
  fasta_filename <- system.file(
    "extdata", "prodigal.fasta", package = "prodigalr"
  )
  filenames <- create_prodigal_files(fasta_filename)
  expect_equal(5, length(filenames))
}
}
\seealso{
use \link{run_prodigal} to received
also the parsed output
}
\author{
Richèl J.C. Bilderbeek
}
