% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_prodigal_options.R
\name{create_prodigal_options}
\alias{create_prodigal_options}
\title{Create the Prodigal options}
\usage{
create_prodigal_options(
  fasta_filename,
  folder_name = get_default_prodigal_folder(),
  output_filename = tempfile("output_"),
  trans_filename = tempfile("trans_"),
  nuc_filename = tempfile("nuc_"),
  start_filename = tempfile("start_")
)
}
\arguments{
\item{fasta_filename}{path to a FASTA file}

\item{folder_name}{superfolder of Prodigal.
The superfolder's name is \code{/home/[user_name]/.local/share}
by default, as can be obtained by
\link{get_default_prodigal_folder}}

\item{output_filename}{name of the file to write the output to}

\item{trans_filename}{name of the file to write the protein translations to}

\item{nuc_filename}{name of the file to write
the nucleotide sequences of genes to}

\item{start_filename}{name of the file to write
all potential genes (with scores) to}
}
\value{
the Prodigal options
}
\description{
Create the Prodigal options
}
\author{
Richèl J.C. Bilderbeek
}
