% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prodigal_version.R
\name{get_prodigal_version}
\alias{get_prodigal_version}
\title{Get the Prodigal version}
\usage{
get_prodigal_version(folder_name = get_default_prodigal_folder())
}
\arguments{
\item{folder_name}{superfolder of Prodigal.
The superfolder's name is \code{/home/[user_name]/.local/share}
by default, as can be obtained by
\link{get_default_prodigal_folder}}
}
\value{
the Prodigal version as a string, for example \code{2.6.3}
}
\description{
Get the Prodigal version
}
\examples{
library(testthat)

if (is_prodigal_installed()) {
  version <- get_prodigal_version()
  expect_true(nchar(version) >= 1)
}
}
