% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_prodigal.R
\name{install_prodigal}
\alias{install_prodigal}
\title{Install Prodigal to a local folder}
\usage{
install_prodigal(folder_name = get_default_prodigal_folder())
}
\arguments{
\item{folder_name}{superfolder of Prodigal.
The superfolder's name is \code{/home/[user_name]/.local/share}
by default, as can be obtained by
\link{get_default_prodigal_folder}}
}
\description{
Install Prodigal to a local folder
}
\examples{
library(testthat)

if (1 == 2) {
  install_prodigal()
  expect_true(is_prodigal_installed())
}
}
\author{
Richèl J.C. Bilderbeek
}
