% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_prodigal_installed.R
\name{is_prodigal_installed}
\alias{is_prodigal_installed}
\title{Measure if Prodigal is installed locally}
\usage{
is_prodigal_installed(folder_name = get_default_prodigal_folder())
}
\arguments{
\item{folder_name}{superfolder of Prodigal.
The superfolder's name is \code{/home/[user_name]/.local/share}
by default, as can be obtained by
\link{get_default_prodigal_folder}}
}
\value{
TRUE is Prodigal is installed locally,
FALSE otherwise
}
\description{
Measure if Prodigal is installed locally
}
\examples{
library(testthat)

is_installed <- is_prodigal_installed()
expect_true(is_installed == TRUE || is_installed == FALSE)
}
\author{
Richèl J.C. Bilderbeek
}
