% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prodigalr.R
\docType{package}
\name{prodigalr}
\alias{prodigalr}
\title{prodigalr: call Prodigal from R}
\description{
It is tricky to predict a proteome from a genome.
'Prodigal' is a tool that allows one to do so.
This package allows one to call 'Prodigal' from R.
}
\examples{
library(testthat)

if (1 == 2 && is_prodigal_installed()) {
  fasta_filename <- system.file("extdata", "prodigal.fasta", package = "prodigal")
  locatome <- run_prodigal(fasta_filename)

  # Show as text
  cat(locatome, sep = "\n")

  # Show as data frame
  print(locatome_to_df(locatome))

  # Show as plot
  plot_locatome(locatome)
}
}
