% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_prodigal.R
\name{run_prodigal}
\alias{run_prodigal}
\title{Runs Prodigal and returns the parsed results}
\usage{
run_prodigal(
  fasta_filename,
  folder_name = get_default_prodigal_folder(),
  temp_folder_name = tempfile(pattern = "pureseqt_")
)
}
\arguments{
\item{fasta_filename}{path to a FASTA file}

\item{folder_name}{superfolder of Prodigal.
The superfolder's name is \code{/home/[user_name]/.local/share}
by default, as can be obtained by
\link{get_default_prodigal_folder}}

\item{temp_folder_name}{path of a temporary folder.
The folder does not need to exist.
Files that are out in this folder are not automatically
deleted, which is not a problem, as the default
path given by \link{tempdir} is automatically cleaned
by the operating system}
}
\value{
full path to the files created
}
\description{
Runs Prodigal and returns the parsed results
}
\examples{
library(testthat)

if (1 == 2 && is_prodigal_installed()) {
  fasta_filename <- system.file(
    "extdata", "prodigal.fasta", package = "prodigal"
  )
  locatome <- run_prodigal(fasta_filename)
  cat(locatome, sep = "\n")
}
}
\seealso{
use \link{create_prodigal_files} to
only create the Prodigal output files
}
\author{
Richèl J.C. Bilderbeek
}
