% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uninstall_prodigal.R
\name{uninstall_prodigal}
\alias{uninstall_prodigal}
\title{Uninstall Prodigal}
\usage{
uninstall_prodigal(folder_name = get_default_prodigal_folder())
}
\arguments{
\item{folder_name}{name of the folder
where the Prodigal files are installed.
The name of the Prodigal binary file will be at
\code{[folder_name]/Prodigal}}
}
\description{
Uninstall Prodigal
}
\author{
Richèl J.C. Bilderbeek
}
