% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_prodigal_files.R
\name{create_prodigal_files}
\alias{create_prodigal_files}
\title{Create the Prodigal output files, by running Prodigal}
\usage{
create_prodigal_files(prodigal_options)
}
\arguments{
\item{prodigal_options}{options to run Prodigal,
as created by \link{create_prodigal_options}}
}
\value{
nothing, will create the files as specified in
the \code{prodigal_options}
}
\description{
Create the Prodigal output files, by running Prodigal
}
\examples{
library(testthat)

if (is_prodigal_installed()) {
  prodigal_options <- create_prodigal_options(
    fasta_filename = system.file(
      "extdata", "covid19.fasta", package = "prodigalr"
    )
  )
  expect_true(are_prodigal_options(prodigal_options))
  expect_silent(create_prodigal_files(prodigal_options))
  expect_true(file.exists(prodigal_options$output_filename))
  expect_true(file.exists(prodigal_options$trans_filename))
  expect_true(file.exists(prodigal_options$nuc_filename))
  expect_true(file.exists(prodigal_options$start_filename))
}
}
\seealso{
use \link{run_prodigal} to received
also the parsed output
}
\author{
Richèl J.C. Bilderbeek
}
